file_ops and env
