+++ tests/src/utils.rs
@@ -67,28 +67,6 @@ fn format_time() -> String {
     format!("{:02}:{:02}:{:02}", now.hour(), now.minute(), now.second())
 }
 
-#[derive(Debug)]
-pub enum FileOps {
-    Python,
-    Rust,
-    GoDir21,
-    GoDir22,
-    GoDir23,
-    GoStatfs,
-}
-
-#[derive(Debug)]
-pub enum EnvApp {
-    Go21,
-    Go22,
-    Go23,
-    Bash,
-    BashInclude,
-    BashExclude,
-    NodeInclude,
-    NodeExclude,
-}
-
 pub struct TestProcess {
     pub child: Child,
     stderr_data: Arc<RwLock<String>>,
@@ -396,56 +374,6 @@ impl TestProcess {
             .unwrap();
         println!("Started application.");
         TestProcess::from_child(child, None)
-    }
-}
-
-impl FileOps {
-    pub fn command(&self) -> Vec<&str> {
-        match self {
-            FileOps::Python => {
-                vec!["python3", "-B", "-m", "unittest", "-f", "python-e2e/ops.py"]
-            }
-            FileOps::Rust => vec!["../target/debug/rust-e2e-fileops"],
-            FileOps::GoDir21 => vec!["go-e2e-dir/21.go_test_app"],
-            FileOps::GoDir22 => vec!["go-e2e-dir/22.go_test_app"],
-            FileOps::GoDir23 => vec!["go-e2e-dir/23.go_test_app"],
-            FileOps::GoStatfs => vec!["go-e2e-statfs/23.go_test_app"],
-        }
-    }
-
-    pub async fn assert(&self, process: TestProcess) {
-        if let FileOps::Python = self {
-            process.assert_python_fileops_stderr().await
-        }
-    }
-}
-
-impl EnvApp {
-    pub fn command(&self) -> Vec<&str> {
-        match self {
-            Self::Go21 => vec!["go-e2e-env/21.go_test_app"],
-            Self::Go22 => vec!["go-e2e-env/22.go_test_app"],
-            Self::Go23 => vec!["go-e2e-env/23.go_test_app"],
-            Self::Bash => vec!["bash", "bash-e2e/env.sh"],
-            Self::BashInclude => vec!["bash", "bash-e2e/env.sh", "include"],
-            Self::BashExclude => vec!["bash", "bash-e2e/env.sh", "exclude"],
-            Self::NodeInclude => vec![
-                "node",
-                "node-e2e/remote_env/test_remote_env_vars_include_works.mjs",
-            ],
-            Self::NodeExclude => vec![
-                "node",
-                "node-e2e/remote_env/test_remote_env_vars_exclude_works.mjs",
-            ],
-        }
-    }
-
-    pub fn mirrord_args(&self) -> Option<Vec<&str>> {
-        match self {
-            Self::BashInclude | Self::NodeInclude => Some(vec!["-s", "MIRRORD_FAKE_VAR_FIRST"]),
-            Self::BashExclude | Self::NodeExclude => Some(vec!["-x", "MIRRORD_FAKE_VAR_FIRST"]),
-            Self::Go21 | Self::Go22 | Self::Go23 | Self::Bash => None,
-        }
     }
 }
